requirements.txt
Augmentor==0.2.12
Keras_Preprocessing==1.1.2
matplotlib==3.5.1
maxflow==0.0.1
numpy==1.22.4
opencv_python==4.5.5.64
opencv_python_headless==4.5.5.64
pandas==1.4.2
Pillow==9.0.1
Pillow==9.5.0
pytablewriter==0.64.2
pytest==7.1.2
PyYAML==6.0
PyYAML==6.0
qudida==0.0.4
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.8.1
setuptools==61.2.0
SimpleITK==2.1.1
SimpleITK==2.2.1
solt==0.1.9
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
